+++ src/main/java/org/notificationplatform/notification/repositories/NotificationEventLogsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NotificationEventLogsRepository extends JpaRepository<NotificationEventLogs, Integer> {
+public interface NotificationEventLogsRepository extends JpaRepository<NotificationEventLogs, Long> {
 
 }
